feat: ajustando homepage para base_view
